[nasa/nos3#319] move classes to own directory
