+++ cfg/gui/classes/add_dialog.py
