+++ cfg/gui/igniter_entrypoint.py
@@ -1,4 +1,4 @@
-from nos3_gui import NOS3ConfigGUI
+from classes.nos3_gui import NOS3ConfigGUI
 
 
 if __name__ == "__main__":
